fix (date problem, decimal cost, prices, rebook extra service when canceled, hours date numbers field, disable book btn after booking)

+++ database/migrations/2021_11_19_212440_create_places_table.php
@@ -19,7 +19,7 @@ class CreatePlacesTable extends Migration
             $table->text('title')->unique();
             $table->text('plc_type');
             $table->text('status');
-            $table->integer('price');
+            $table->decimal('price', 15, 2);
             $table->integer('capacity');
             $table->longText('description');
             $table->string('image')->nullable();
